Refactor to use uint8_t instead of byte

+++ src/OBD2_KLine.h
@@ -13,62 +13,62 @@
 #endif
 
 // ==== OBD2 Mods ====
-const byte init_OBD = 0x81;              // Init ISO14230
-const byte read_LiveData = 0x01;         // Read Live Data
-const byte read_FreezeFrame = 0x02;      // Read Freeze Frame Data
-const byte read_storedDTCs = 0x03;       // Read Stored Troubleshoot Codes
-const byte clear_DTCs = 0x04;            // Clear Troubleshoot Codes
-const byte test_OxygenSensors = 0x05;    // Test Oxygen Sensors
-const byte component_Monitoring = 0x06;  // Component Monitoring
-const byte read_pendingDTCs = 0x07;      // Read Pending Troubleshoot Codes
-const byte read_VehicleInfo = 0x09;      // Read Vehicle Info
-const byte SUPPORTED_PIDS_1_20 = 0x00;
-const byte SUPPORTED_PIDS_21_40 = 0x20;
-const byte SUPPORTED_PIDS_41_60 = 0x40;
+const uint8_t init_OBD = 0x81;              // Init ISO14230
+const uint8_t read_LiveData = 0x01;         // Read Live Data
+const uint8_t read_FreezeFrame = 0x02;      // Read Freeze Frame Data
+const uint8_t read_storedDTCs = 0x03;       // Read Stored Troubleshoot Codes
+const uint8_t clear_DTCs = 0x04;            // Clear Troubleshoot Codes
+const uint8_t test_OxygenSensors = 0x05;    // Test Oxygen Sensors
+const uint8_t component_Monitoring = 0x06;  // Component Monitoring
+const uint8_t read_pendingDTCs = 0x07;      // Read Pending Troubleshoot Codes
+const uint8_t read_VehicleInfo = 0x09;      // Read Vehicle Info
+const uint8_t SUPPORTED_PIDS_1_20 = 0x00;
+const uint8_t SUPPORTED_PIDS_21_40 = 0x20;
+const uint8_t SUPPORTED_PIDS_41_60 = 0x40;
 
-const byte supported_VehicleInfo = 0x00;  // Read Supported Vehicle Info
-const byte read_VIN_Count = 0x01;         // Read VIN Count
-const byte read_VIN = 0x02;               // Read VIN
-const byte read_ID_Length = 0x03;         // Read Calibration ID Length
-const byte read_ID = 0x04;                // Read Calibration ID
-const byte read_ID_Num_Length = 0x05;     // Read Calibration ID Number Length
-const byte read_ID_Num = 0x06;            // Read Calibration ID Number
+const uint8_t supported_VehicleInfo = 0x00;  // Read Supported Vehicle Info
+const uint8_t read_VIN_Count = 0x01;         // Read VIN Count
+const uint8_t read_VIN = 0x02;               // Read VIN
+const uint8_t read_ID_Length = 0x03;         // Read Calibration ID Length
+const uint8_t read_ID = 0x04;                // Read Calibration ID
+const uint8_t read_ID_Num_Length = 0x05;     // Read Calibration ID Number Length
+const uint8_t read_ID_Num = 0x06;            // Read Calibration ID Number
 
 class OBD2_KLine {
  public:
   //VehicleData car;
-  OBD2_KLine(SerialType &serialStream, long baudRate, uint8_t rxPin, uint8_t txPin);
+  OBD2_KLine(SerialType &serialStream, uint32_t baudRate, uint8_t rxPin, uint8_t txPin);
 
   void setDebug(Stream &serial);
   void setSerial(bool enabled);
   bool initOBD2();
   bool trySlowInit();
   bool tryFastInit();
-  void writeData(byte mode, byte pid);
-  void writeRawData(const byte *dataArray, int length);
-  int readData();
+  void writeData(uint8_t mode, uint8_t pid);
+  void writeRawData(const uint8_t *dataArray, uint8_t length);
+  uint8_t readData();
   void send5baud(uint8_t data);
 
-  float getPID(byte mode, byte pid);
-  float getLiveData(byte pid);
-  float getFreezeFrame(byte pid);
+  float getPID(uint8_t mode, uint8_t pid);
+  float getLiveData(uint8_t pid);
+  float getFreezeFrame(uint8_t pid);
 
-  int readDTCs(byte mode);
-  int readStoredDTCs();
-  int readPendingDTCs();
-  String getStoredDTC(int index);
-  String getPendingDTC(int index);
+  uint8_t readDTCs(uint8_t mode);
+  uint8_t readStoredDTCs();
+  uint8_t readPendingDTCs();
+  String getStoredDTC(uint8_t index);
+  String getPendingDTC(uint8_t index);
 
   bool clearDTC();
 
-  String getVehicleInfo(byte pid);
+  String getVehicleInfo(uint8_t pid);
 
-  int readSupportedLiveData();
-  int readSupportedFreezeFrame();
-  int readSupportedComponentMonitoring();
-  int readSupportedVehicleInfo();
-  int readSupportedData(byte mode);
-  byte getSupportedData(byte mode, int index);
+  uint8_t readSupportedLiveData();
+  uint8_t readSupportedFreezeFrame();
+  uint8_t readSupportedComponentMonitoring();
+  uint8_t readSupportedVehicleInfo();
+  uint8_t readSupportedData(uint8_t mode);
+  uint8_t getSupportedData(uint8_t mode, uint8_t index);
 
   void setWriteDelay(uint16_t delay);
   void setDataRequestInterval(uint16_t interval);
@@ -76,7 +76,7 @@ class OBD2_KLine {
 
  private:
   SerialType *_serial;
-  long _baudRate;
+  uint32_t _baudRate;
   uint8_t _rxPin;
   uint8_t _txPin;
   Stream *_debugSerial = nullptr;  // Debug serial port
